Struct octocat_rs::github::client_builder::ClientBuilder

pub struct ClientBuilder<T>where
    T: Debug + EventHandler<GitHubClient = Client<T>> + Send + Sync + 'static,{ /* private fields */ }

A builder for Client

Implementations

impl<T> ClientBuilder<T>where T: Debug + EventHandler<GitHubClient = Client<T>> + Send + Sync + 'static,

pub fn new() -> Self

Creates a new ClientBuilder

pub fn event_handler(self, event_handler: T) -> Self

Adds an EventHandler to the current builder.

pub fn payload_size(self, size: u64) -> Self

Sets the maximum payload size that the listener can receive from GitHub in MiB. Default: 8.

pub fn user_agent<V: Into<String>>(self, user_agent: V) -> Self

Sets a custom user agent for your application. Default is “Octocat-rs”.

See also: HttpClient::set_ua

pub fn credentials_file<P: AsRef<Path>>(self, file: P) -> Self

Adds an Authorization instance to the current builder using input from a file.

pub fn credentials_env_var<K: AsRef<OsStr>>( self, username_var: K, token_var: K ) -> Self

Adds an Authorization instance to the current builder using input from an environment variable.

pub fn personal_auth<V: Into<String>>(self, username: V, token: V) -> Self

Adds an Authorization instance to the current builder.

pub fn build(self) -> Result<Client<T>>

Builds the current builder. In other words, this turns a ClientBuilder into a Client. Requires a handler to be set.

impl ClientBuilder<DefaultEventHandler>

pub fn build_unconfigured() -> Client<DefaultEventHandler>

Returns the default implementation of Client

pub fn build_no_handler(self) -> Result<Client<DefaultEventHandler>>

For building the current builder without setting a handler.

Requires T to be set to DefaultEventHandler.