Module octocat_rs::github::handler

Contains the EventHandler trait and its default implementation (DefaultEventHandler).

Structs

DefaultEventHandler

Traits

EventHandler

An event handler that is used in all clients. For end users, an implementer of this trait is passed to a ClientBuilder instance when creating the client in your main function.