Struct octocat_rs::github::http::HttpClient

pub struct HttpClient { /* private fields */ }

An implementer of the Requester trait. This is all most users will need, however it may be helpful to look at the trait implementation details here if you’re writing your own implementation.

Implementations

impl HttpClient

pub fn new(auth: Option<Authorization>, user_agent: Option<String>) -> Self

Creates a new HttpClient.

pub fn new_none() -> Self

Creates a new HttpClient with no authorization.

pub fn set_auth(&mut self, auth: Authorization)

Updates the authorization used by the current client.

pub fn set_ua(&mut self, user_agent: &str)

Set the user agent used by the current client.