Trait octocat_rs::github::handler::EventHandler

pub trait EventHandler {
    type Message: Debug + Send;
    type GitHubClient: GitHubClient + Send + Sync;

    fn listener_port(&self) -> u16 { ... }
    fn route(&self) -> &'static str { ... }
    fn message<'life0, 'async_trait>(
        &'life0 self, 
        message: Self::Message
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn app_authorization_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        app_authorization_event: AppAuthorizationEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn installation_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        installation_event: InstallationEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn installation_repositories_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        installation_repositories_event: InstallationRepositoriesEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn push_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        push_event: PushEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn commit_comment_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        commit_comment_event: CommitCommentEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn release_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        release_event: ReleaseEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn repository_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        repository_event: RepositoryEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn organization_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        organization_event: OrganizationEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn team_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        team_event: TeamEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn team_add_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        team_add_event: TeamAddEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn project_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        project_event: ProjectEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn project_card_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        project_card_event: ProjectCardEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn project_column_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        project_column_event: ProjectColumnEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn meta_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        meta_event: MetaEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn package_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        package_event: PackageEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn ping_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        ping_event: PingEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn sponsorship_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        sponsorship_event: SponsorshipEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn marketplace_purchase_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        marketplace_purchase_event: MarketplacePurchaseEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn repository_dispatch_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        repository_dispatch_event: RepositoryDispatchEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn repository_import_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        repository_import_event: RepositoryImportEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn repository_vulnerability_alert<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        repository_vulnerability_alert: RepositoryVulnerabilityAlertEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn secret_scanning_alert<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        secret_scanning_alert: SecretScanningAlertEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn status_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        status_event: StatusEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn label_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        label_event: LabelEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn discussion_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        discussion_event: DiscussionEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn discussion_comment_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        discussion_comment_event: DiscussionCommentEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn branch_protection_rule_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        branch_protection_rule_event: BranchProtectionRuleEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn check_suite_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        check_suite_event: CheckSuiteEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn code_scanning_alert<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        code_scanning_alert: CodeScanningAlertEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn deployment_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        deployment_event: DeploymentEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn deployment_status_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        deployment_status_event: DeploymentStatusEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn page_build_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        page_build_event: PageBuildEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn workflow_dispatch_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        workflow_dispatch_event: WorkflowDispatchEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn deploy_key_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        deploy_key_event: DeployKeyEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn gollum_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        gollum_event: GollumEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn membership_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        membership_event: MembershipEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn org_block_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        org_block_event: OrgBlockEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn member_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        member_event: MemberEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn public_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        public_event: PublicEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn milestone_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        milestone_event: MilestoneEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn tag_created<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        create_event: CreateEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn tag_deleted<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        delete_event: DeleteEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn star_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        star_event: StarEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn security_advisory_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        security_advisory_event: SecurityAdvisoryEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn watch_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        watch_event: WatchEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn repository_forked<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        fork_event: ForkEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn pull_request_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        pull_request_event: PullRequestEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn pull_request_review_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        pull_request_review_event: PullRequestReviewEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn pull_request_review_comment_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        pull_request_review_comment_event: PullRequestReviewCommentEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn workflow_run<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        workflow_run: WorkflowRunEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn workflow_job<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        workflow_job: WorkflowJobEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn check_run<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        check_run: CheckRunEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn issue_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        issue_event: IssueEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn issue_comment_event<'life0, 'async_trait>(
        &'life0 self, 
        github_client: Arc<Self::GitHubClient>, 
        issue_comment_event: IssueCommentEvent
    ) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

}

An event handler that is used in all clients. For end users, an implementer of this trait is passed to a ClientBuilder instance when creating the client in your main function.

Required Associated Types

type Message: Debug + Send

type GitHubClient: GitHubClient + Send + Sync

Provided Methods

fn listener_port(&self) -> u16

Utility function for setting the port used by the webhook.

fn route(&self) -> &'static str

The route at which the listener should listen for payloads from GitHub.

fn message<'life0, 'async_trait>(
&'life0 self,
message: Self::Message
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

fn app_authorization_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
app_authorization_event: AppAuthorizationEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Someone revokes their authorization of a GitHub App

fn installation_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
installation_event: InstallationEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a GitHub App installation

fn installation_repositories_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
installation_repositories_event: InstallationRepositoriesEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to repositories being added to a GitHub App installation

fn push_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
push_event: PushEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Commit pushed to a repository

fn commit_comment_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
commit_comment_event: CommitCommentEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Comment added to a repository commit

fn release_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
release_event: ReleaseEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Release created

fn repository_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
repository_event: RepositoryEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Repository-related activity

fn organization_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
organization_event: OrganizationEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Organization-related activity

fn team_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
team_event: TeamEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to an organization’s team

fn team_add_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
team_add_event: TeamAddEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

A repository is added to a team

fn project_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
project_event: ProjectEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to project boards

fn project_card_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
project_card_event: ProjectCardEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to project cards

fn project_column_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
project_column_event: ProjectColumnEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to columns in a project board

fn meta_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
meta_event: MetaEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

The webhook this event is configured on was deleted. This event will only listen for changes to the particular hook the event is installed on

fn package_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
package_event: PackageEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to GitHub Packages

fn ping_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
ping_event: PingEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

New webhook created

fn sponsorship_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
sponsorship_event: SponsorshipEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a sponsorship listing

fn marketplace_purchase_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
marketplace_purchase_event: MarketplacePurchaseEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a GitHub Marketplace purchase

fn repository_dispatch_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
repository_dispatch_event: RepositoryDispatchEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

GitHub App sends a POST request to the “create a repository dispatch event” endpoint

fn repository_import_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
repository_import_event: RepositoryImportEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a repository being imported to GitHub

fn repository_vulnerability_alert<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
repository_vulnerability_alert: RepositoryVulnerabilityAlertEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to security vulnerability alerts in a repository

fn secret_scanning_alert<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
secret_scanning_alert: SecretScanningAlertEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to secret scanning alerts in a repository

fn status_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
status_event: StatusEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Status of a Git commit changed

fn label_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
label_event: LabelEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a label.

fn discussion_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
discussion_event: DiscussionEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a discussion

fn discussion_comment_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
discussion_comment_event: DiscussionCommentEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a comment in a discussion

fn branch_protection_rule_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
branch_protection_rule_event: BranchProtectionRuleEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a branch protection rule

fn check_suite_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
check_suite_event: CheckSuiteEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Check suite activity has occurred

fn code_scanning_alert<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
code_scanning_alert: CodeScanningAlertEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to code scanning alerts in a repository

fn deployment_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
deployment_event: DeploymentEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

A deployment is created

fn deployment_status_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
deployment_status_event: DeploymentStatusEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

A deployment is created

fn page_build_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
page_build_event: PageBuildEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Represents an attempted build of a GitHub Pages site (successful or not)

fn workflow_dispatch_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
workflow_dispatch_event: WorkflowDispatchEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Someone triggers a workflow run on GitHub or sends a POST request to the “create a workflow dispatch event” endpoint

fn deploy_key_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
deploy_key_event: DeployKeyEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Deploy key added or removed from a repository

fn gollum_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
gollum_event: GollumEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

A wiki page is created or updated

fn membership_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
membership_event: MembershipEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to team membership

fn org_block_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
org_block_event: OrgBlockEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to a user being blocked in an organization

fn member_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
member_event: MemberEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Event related to repository collaborators

fn public_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
public_event: PublicEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Repository made public

fn milestone_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
milestone_event: MilestoneEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Event related to repository milestones

fn tag_created<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
create_event: CreateEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Git branch or tag created

fn tag_deleted<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
delete_event: DeleteEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Git branch or tag deleted

fn star_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
star_event: StarEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Repository receives a star

fn security_advisory_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
security_advisory_event: SecurityAdvisoryEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

fn watch_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
watch_event: WatchEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Someone begins watching a repository

fn repository_forked<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
fork_event: ForkEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Repository is forked

fn pull_request_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
pull_request_event: PullRequestEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to pull requests

fn pull_request_review_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
pull_request_review_event: PullRequestReviewEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to pull request reviews

fn pull_request_review_comment_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
pull_request_review_comment_event: PullRequestReviewCommentEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to pull request review comments in the pull request’s unified diff

fn workflow_run<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
workflow_run: WorkflowRunEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

A GitHub Actions workflow run is requested or completed

fn workflow_job<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
workflow_job: WorkflowJobEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

A GitHub Actions workflow job has been queued, is in progress, or has been completed on a repository

fn check_run<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
check_run: CheckRunEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Check run activity has occurred

fn issue_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
issue_event: IssueEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to an issue

fn issue_comment_event<'life0, 'async_trait>(
&'life0 self,
github_client: Arc<Self::GitHubClient>,
issue_comment_event: IssueCommentEvent
) -> Pin<Box<dyn Future<Output = Command<Self::Message>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Activity related to an issue or pull request comment